Sync 1.44 r1033
[01033 ]: w92 include fixes, doomcincl.h
Added the forgettn important doomcincl.h from svn1058

+++ SRC/tables.c
@@ -1,7 +1,7 @@
 // Emacs style mode select   -*- C++ -*- 
 //-----------------------------------------------------------------------------
 //
-// $Id: tables.c 890 2011-12-26 18:25:54Z wesleyjohnson $
+// $Id: tables.c 1035 2013-08-14 00:38:40Z wesleyjohnson $
 //
 // Copyright (C) 1993-1996 by id Software, Inc.
 // Portions Copyright (C) 1998-2000 by DooM Legacy Team.
@@ -18,9 +18,9 @@
 //
 //
 // $Log: tables.c,v $
-// Revision 1.2  2000/02/27 00:42:11  hurdler
-// fix CR+LF problem
+// Include: DOS DJGPP Fixes
 //
+// Revision 1.2  2000/02/27 00:42:11  hurdler
 // Revision 1.1.1.1  2000/02/22 20:32:32  hurdler
 // Initial import into CVS (v1.29 pr3)
 //
@@ -81,8 +81,8 @@ const angle_t ANGLE_60  = 0x80000000 / 3;
     int SlopeDiv ( unsigned      num,
                    unsigned      den)
     {
-    ULONG  ans;
-    INT64  t;
+    uint32_t ans;
+    int64_t  t;
 
     if (den < 512)
         return SLOPERANGE;
@@ -118,7 +118,7 @@ const angle_t ANGLE_60  = 0x80000000 / 3;
 
 // The finetangentgent[angle+FINEANGLES/4] table
 // holds the fixed_t tangent values for view angles,
-// ranging from -FIXED_MAX to 0 to FIXED_MAX (ranging from MININT to 0 to MAXINT.)
+// ranging from -FIXED_MAX to 0 to FIXED_MAX
 // ([WDJ] actually more like -FIXED_MAX/12.56 to FIXED_MAX/12.56).
 // fixed_t              finetangent[FINEANGLES/2];
